у части View вынес формирование разметки отдельно
